Move comments into INaturalModule
